fix for model evaluation

+++ src/engineering/LitCallbacks.py
@@ -31,7 +31,7 @@ class PSDCallbacks:
 class LoggingCallback(Callback):
     def on_validation_epoch_end(self, trainer, pl_module):
         if hasattr(pl_module, "confusion_matrix"):
-            pl_module.logger.experiment.add_figure("validation/confusion_matrix", plot_confusion_matrix(pl_module.confusion_matrix.detach().cpu().numpy(),
+            pl_module.logger.experiment.add_figure("validation/confusion_matrix", plot_confusion_matrix(pl_module.confusion_matrix.detach().cpu().numpy().astype(int),
                                                                                         pl_module.config.system_config.type_names,
                                                                                         normalize=True))
             pl_module.confusion_matrix = zeros(pl_module.confusion_matrix.shape, device=pl_module.device)
